pyproject.toml
[project]
name = "hormoziai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gradio>=5.43.1",
    "groq>=0.31.0",
    "python-dotenv>=1.1.1",
]
